+++ packages/beta/src/__tests__/cogniteClient.int.spec.ts
@@ -13,7 +13,8 @@ describe('beta integration', () => {
     expect(response.items.length).toBeGreaterThan(0);
   });
   test('raw get assets', async () => {
-    const response = await client.get('/api/v1/projects/cognitesdk-js/assets');
+    const project = process.env.COGNITE_PROJECT as string;
+    const response = await client.get(`/api/v1/projects/${project}/assets`);
     expect(response.data).toHaveProperty('items');
   });
 });