+++ packages/stable/src/__tests__/testUtils.ts
@@ -18,26 +18,32 @@ export const {
 } = TestUtils;
 
 export function setupClient(baseUrl: string = Constants.BASE_URL) {
-  return new CogniteClient({ appId: 'JS SDK integration tests', baseUrl });
+  return new CogniteClient({
+    appId: 'JS SDK integration tests',
+    project: 'unit-test',
+    getToken: () => Promise.reject(new Error('not logged in')),
+    baseUrl,
+  });
 }
 
 export function setupLoggedInClient() {
-  jest.setTimeout(60 * 1000);
-  const client = setupClient();
-  client.loginWithApiKey({
+  return new CogniteClient({
+    appId: 'JS SDK integration tests',
     project: process.env.COGNITE_PROJECT as string,
-    apiKey: process.env.COGNITE_CREDENTIALS as string,
+    getToken: () => Promise.resolve(process.env.COGNITE_CREDENTIALS as string),
+    apiKeyMode: true,
+    baseUrl: Constants.BASE_URL,
   });
-  return client;
 }
 
 export function setupMockableClient() {
-  const client = setupClient(mockBaseUrl);
-  client.loginWithApiKey({
-    project,
-    apiKey,
+  return new CogniteClient({
+    appId: 'JS SDK integration tests',
+    project: process.env.COGNITE_PROJECT as string,
+    getToken: () => Promise.resolve(process.env.COGNITE_CREDENTIALS as string),
+    apiKeyMode: true,
+    baseUrl: mockBaseUrl,
   });
-  return client;
 }
 
 export const getFileCreateArgs = (