+++ packages/playground/CHANGELOG.md
@@ -3,6 +3,81 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [3.1.2](https://github.com/cognitedata/cognite-sdk-js/compare/@cognite/sdk-playground@3.1.1...@cognite/sdk-playground@3.1.2) (2021-10-12)
+
+**Note:** Version bump only for package @cognite/sdk-playground
+
+
+
+
+
+## [3.1.1](https://github.com/cognitedata/cognite-sdk-js/compare/@cognite/sdk-playground@3.1.0...@cognite/sdk-playground@3.1.1) (2021-10-12)
+
+
+### Bug Fixes
+
+* **documents:** pipeline endpoints ([#701](https://github.com/cognitedata/cognite-sdk-js/issues/701)) ([c5930bb](https://github.com/cognitedata/cognite-sdk-js/commit/c5930bb4bde38cc8d93b3cbba313a5186ba4335f))
+
+
+
+
+
+# [3.1.0](https://github.com/cognitedata/cognite-sdk-js/compare/@cognite/sdk-playground@3.0.0...@cognite/sdk-playground@3.1.0) (2021-10-12)
+
+
+### Features
+
+* **documents:** support string array ([#702](https://github.com/cognitedata/cognite-sdk-js/issues/702)) ([cb7b5e3](https://github.com/cognitedata/cognite-sdk-js/commit/cb7b5e3c304ee5dbe7753f6d271c6f3ed05fcc2e))
+
+
+
+
+
+# [3.0.0](https://github.com/cognitedata/cognite-sdk-js/compare/@cognite/sdk-playground@2.7.1...@cognite/sdk-playground@3.0.0) (2021-10-12)
+
+
+### Features
+
+* **auth:** re-release auth patch ([#700](https://github.com/cognitedata/cognite-sdk-js/issues/700)) ([a53c40d](https://github.com/cognitedata/cognite-sdk-js/commit/a53c40ddd7eca5d2dee9149f5df0b2e533d19575))
+
+
+### BREAKING CHANGES
+
+* **auth:** release v6
+
+re-release (revert reversion) of "feat(core): move authentication out of CogniteClient"
+https://github.com/cognitedata/cognite-sdk-js/pull/687
+
+This reverts commit 72e1ecb61603e0ac3926124c26f4e009df88f020.
+
+Co-authored-by: Vegard Økland <vegard.okland@cognite.com>
+
+
+
+
+
+## [2.7.1](https://github.com/cognitedata/cognite-sdk-js/compare/@cognite/sdk-playground@2.7.0...@cognite/sdk-playground@2.7.1) (2021-10-12)
+
+
+### Bug Fixes
+
+* **release:** undo major version release without major version bump ([#697](https://github.com/cognitedata/cognite-sdk-js/issues/697)) ([72e1ecb](https://github.com/cognitedata/cognite-sdk-js/commit/72e1ecb61603e0ac3926124c26f4e009df88f020)), closes [#687](https://github.com/cognitedata/cognite-sdk-js/issues/687)
+
+
+
+
+
+# [2.7.0](https://github.com/cognitedata/cognite-sdk-js/compare/@cognite/sdk-playground@2.6.3...@cognite/sdk-playground@2.7.0) (2021-10-12)
+
+
+### Features
+
+* **core:** move authentication out of CogniteClient ([#687](https://github.com/cognitedata/cognite-sdk-js/issues/687)) ([879ed31](https://github.com/cognitedata/cognite-sdk-js/commit/879ed31d05dd6d6f4b691b99eaca5fa7363e96e6))
+
+
+
+
+
 ## [2.6.3](https://github.com/cognitedata/cognite-sdk-js/compare/@cognite/sdk-playground@2.6.2...@cognite/sdk-playground@2.6.3) (2021-10-07)
 
 **Note:** Version bump only for package @cognite/sdk-playground